Backend/requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic
python-dotenv
psycopg2-binary
SQLAlchemy